modify whole code

+++ nukkad_App/src/main/java/com/code/monks/nukkad/dto/request/CreateCartItemRequestDTO.java
@@ -10,7 +10,7 @@ public class CreateCartItemRequestDTO {
 
     @Data
     public static class CartItemRequest {
-        private int itemId;
+        private Long itemId;
         private int quantity;
         private String unit;
     }
